Applied patches to libgeographic and mrt_cmake_modules
This was done to make it possible to build lanelet2-projection
as it was failing on this step. libgeographic has been renamed
in version 2.0 and the automated build system mrt_cmake_modules
that lanelet2-projection was using failed because of this.

Note that in the current latest version, this issue has been fixed.
This is only applied because the lopsised98 nix-ros-overlay repo
master branch has older packages.

Reference: https://github.com/KIT-MRT/mrt_cmake_modules/commit/15d2ff0f033bbfc786812a0305fe50711a3dbab7